RDCC-320 ensuring emails are made lowercase
