+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/NewUserCreationRequest.java
@@ -38,6 +38,8 @@ public class NewUserCreationRequest {
             @JsonProperty("status") String status,
             @JsonProperty("roles") List<String> roles) {
 
+        if(email != null) {email = email.toLowerCase();}
+
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
