+++ src/main/java/uk/gov/hmcts/reform/professionalapi/controller/request/UserCreationRequest.java
@@ -25,6 +25,7 @@ public class UserCreationRequest {
             @JsonProperty("lastName") String lastName,
             @JsonProperty("email") String email
     ) {
+        if(email != null) {email = email.toLowerCase();}
 
         this.firstName = firstName;
         this.lastName = lastName;
