+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/request/NewUserCreationRequestTest.java
@@ -17,7 +17,7 @@ public class NewUserCreationRequestTest {
     private final Validator validator = Validation.buildDefaultValidatorFactory().getValidator();
 
     @Test
-    public void has_mandatory_fields_specified_not_null() {
+    public void testHasMandatoryFieldsSpecified() {
         NewUserCreationRequest newUserCreationRequest =
                 new NewUserCreationRequest(null, null, null, null, null);
 
@@ -27,7 +27,7 @@ public class NewUserCreationRequestTest {
     }
 
     @Test
-    public void creates_new_user_creation_request_correctly() {
+    public void testCreatesNewUser() {
         List<String> userRoles = new ArrayList<>();
         userRoles.add("pui-user-manager");
 
@@ -42,7 +42,7 @@ public class NewUserCreationRequestTest {
     }
 
     @Test
-    public void does_not_create_new_user_creation_request_when_email_is_not_valid() {
+    public void testDoesNotCreateNewUserWhenEmailIsInvalid() {
         List<String> userRoles = new ArrayList<>();
         userRoles.add("pui-user-manager");
 
@@ -52,6 +52,16 @@ public class NewUserCreationRequestTest {
         Set<ConstraintViolation<NewUserCreationRequest>> violations = validator.validate(newUserCreationRequest);
 
         assertThat(violations.size()).isEqualTo(1);
+    }
+
+    @Test
+    public void testEmailIsMadeLowercase() {
+        List<String> userRoles = new ArrayList<>();
+        userRoles.add("pui-user-manager");
+
+        NewUserCreationRequest newUserCreationRequest =
+                new NewUserCreationRequest("some-name", "some-last-name", "Dummy@EMAIL.com", "PENDING", userRoles);
+        assertThat(newUserCreationRequest.getEmail()).isEqualTo("dummy@email.com");
     }
 
     @Test
