+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/request/controller/request/UserCreationRequestTest.java
@@ -18,7 +18,7 @@ public class UserCreationRequestTest {
 
     private String firstName = "Joe";
     private String lastName = "Bloggs";
-    private String email = "joe.bloggs@email.com";
+    private String email = "joe.BLOGGS@Email.Com";
 
     @Before
     public void setUp() {
@@ -26,7 +26,7 @@ public class UserCreationRequestTest {
     }
 
     @Test
-    public void has_mandatory_fields_specified_not_null() {
+    public void hasMandatoryFieldsSpecifiedNotNull() {
 
         UserCreationRequest userCreationRequest =
                 new UserCreationRequest(null, null, null);
@@ -35,6 +35,11 @@ public class UserCreationRequestTest {
                 validator.validate(userCreationRequest);
 
         assertThat(violations.size()).isEqualTo(3);
+    }
+
+    @Test
+    public void testEmailIsMadeLowercase() {
+        assertThat(userCreationRequest.getEmail()).isEqualTo("joe.bloggs@email.com");
     }
 
     @Test
